# Make ping benchmark bi-directional and add a test