--- a/perfkitbenchmarker/linux_benchmarks/ping_benchmark.py
+++ b/perfkitbenchmarker/linux_benchmarks/ping_benchmark.py
@@ -43,12 +43,15 @@ def GetConfig(user_config):
 
 def Prepare(benchmark_spec):  # pylint: disable=unused-argument
   """Install ping on the target vm.
-
+  Checks that there are exactly two vms specified.
   Args:
     benchmark_spec: The benchmark specification. Contains all data that is
         required to run the benchmark.
   """
-  pass
+  if len(benchmark_spec.vms) != 2:
+    raise ValueError(
+        'Ping benchmark requires exactly two machines, found {0}'
+        .format(len(benchmark_spec.vms)))
 
 
 def Run(benchmark_spec):
@@ -62,17 +65,39 @@ def Run(benchmark_spec):
     A list of sample.Sample objects.
   """
   vms = benchmark_spec.vms
-  if not vms[0].IsReachable(vms[1]):
-    logging.warn('%s is not reachable from %s', vms[1], vms[0])
+  results = []
+  for (sending_vm, receiving_vm) in vms, reversed(vms):
+    results = results + _RunPing(sending_vm,
+                                 receiving_vm,
+                                 receiving_vm.internal_ip,
+                                 'internal')
+  return results
+
+
+def _RunPing(sending_vm, receiving_vm, receiving_ip, ip_type):
+  """Run ping using 'sending_vm' to connect to 'receiving_ip'.
+
+  Args:
+    sending_vm: The VM issuing the ping request.
+    receiving_vm: The VM receiving the ping.  Needed for metadata.
+    receiving_ip: The IP address to be pinged.
+    ip_type: The type of 'receiving_ip' (either 'internal' or 'external')
+  Returns:
+    A list of samples, with one sample for each metric.
+  """
+  if not sending_vm.IsReachable(receiving_vm):
+    logging.warn('%s is not reachable from %s', receiving_vm, sending_vm)
     return []
-  vm = vms[0]
+
   logging.info('Ping results:')
-  ping_cmd = 'ping -c 100 %s' % vms[1].internal_ip
-  stdout, _ = vm.RemoteCommand(ping_cmd, should_log=True)
+  ping_cmd = 'ping -c 100 %s' % receiving_ip
+  stdout, _ = sending_vm.RemoteCommand(ping_cmd, should_log=True)
   stats = re.findall('([0-9]*\\.[0-9]*)', stdout.splitlines()[-1])
   assert len(stats) == len(METRICS), stats
   results = []
-  metadata = {'ip_type': 'internal'}
+  metadata = {'ip_type': ip_type,
+              'receiving_zone': receiving_vm.zone,
+              'sending_zone': sending_vm.zone}
   for i, metric in enumerate(METRICS):
     results.append(sample.Sample(metric, float(stats[i]), 'ms', metadata))
   return results