… tests/data/ping.out