This makes the ping benchmark bi-directional and adds a test.  It also adds `sending_zone` and `receiving_zone` to the metadata.

Old results summary:
```
PING:
  ip_type="internal"
  Min Latency                          36.244000 ms                            
  Average Latency                      36.879000 ms                            
  Max Latency                          89.677000 ms                            
  Latency Std Dev                       5.310000 ms                            
  End to End Runtime                  290.041688 seconds  
```

New results summary:
```
PING:
  ip_type="internal"
  Min Latency                          36.198000 ms                             (receiving_zone="us-east1-b" sending_zone="us-central1-a")
  Average Latency                      37.434000 ms                             (receiving_zone="us-east1-b" sending_zone="us-central1-a")
  Max Latency                         149.869000 ms                             (receiving_zone="us-east1-b" sending_zone="us-central1-a")
  Latency Std Dev                      11.300000 ms                             (receiving_zone="us-east1-b" sending_zone="us-central1-a")
  Min Latency                          36.183000 ms                             (receiving_zone="us-central1-a" sending_zone="us-east1-b")
  Average Latency                      36.308000 ms                             (receiving_zone="us-central1-a" sending_zone="us-east1-b")
  Max Latency                          36.496000 ms                             (receiving_zone="us-central1-a" sending_zone="us-east1-b")
  Latency Std Dev                       0.188000 ms                             (receiving_zone="us-central1-a" sending_zone="us-east1-b")
  End to End Runtime                  409.803674 seconds                       

```