@hildrum 2015-12-01
I had copied the for-loop syntax from the`iperf_benchmark.py` so I modified that benchmark to follow @skschneider's pattern as well.