@skschneider 2015-12-01
A few small comments, otherwise LGTM.